Migrate schedule/submissions tasks to celery
